11455 - refactored topCategories.js datamapping
