+++ src/js/components/recipient/topFive/TopFiveSection.jsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 
-import { recipientCategories as topCategories } from 'dataMapping/recipients/topCategories';
+import { recipientCategories as topCategories } from 'dataMapping/topCategories';
 import TopFiveContainer from 'containers/recipient/topFive/TopFiveContainer';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { SectionHeader } from "data-transparency-ui";
