--- src/js/dataMapping/recipients/topCategories.js
@@ -1,26 +1,0 @@
-/**
- * topCategories.js
- * Created by Kwadwo Opoku-Debrah 07/10/18
- */
-
-export const recipientCategories = [
-    'awarding_agency',
-    'awarding_subagency',
-    'federal_account',
-    'cfda',
-    'naics',
-    'psc',
-    'country',
-    'state_territory'
-];
-
-export const recipientCategoryTitles = {
-    awarding_agency: 'Awarding Agencies',
-    awarding_subagency: 'Awarding Sub-Agencies',
-    federal_account: 'Federal Accounts',
-    cfda: 'Assistance Listings (CFDA Programs)',
-    naics: 'NAICS Codes',
-    psc: 'Product Service Codes',
-    country: 'Countries',
-    state_territory: 'U.S. States or Territories'
-};
