+++ src/js/dataMapping/topCategories.js
@@ -75,5 +75,20 @@ export const categoryTitles = {
     cfda: 'Assistance Listings (CFDA Programs)',
     naics: 'NAICS Codes',
     county: 'Counties',
-    district: 'Congressional Districts'
+    district: 'Congressional Districts',
+    federal_account: 'Federal Accounts',
+    psc: 'Product Service Codes',
+    country: 'Countries',
+    state_territory: 'U.S. States or Territories'
 };
+
+export const recipientCategories = [
+    'awarding_agency',
+    'awarding_subagency',
+    'federal_account',
+    'cfda',
+    'naics',
+    'psc',
+    'country',
+    'state_territory'
+];
